refactor: remove jslint validator
