+++ bundles/com.aptana.editor.js/plugin.xml
@@ -140,10 +140,6 @@
    <extension
          point="com.aptana.build.ui.buildParticipantPreferenceComposite">
       <preferenceComposite
-            buildParticipantId="com.aptana.js.core.JSLintValidator"
-            class="com.aptana.editor.js.validator.JSLintValidatorPreferenceCompositeFactory">
-      </preferenceComposite>
-      <preferenceComposite
             buildParticipantId="com.aptana.js.core.JSParserValidator"
             class="com.aptana.editor.js.validator.JSParserValidatorPreferenceCompositeFactory">
       </preferenceComposite>
