+++ bundles/com.aptana.editor.js/src/com/aptana/editor/js/validator/messages.properties
@@ -1,5 +1,2 @@
-JSLintValidatorPreferenceCompositeFactory_OptionsMsg=Please specify JSLint options using a JSON object.
-JSLintValidatorPreferenceCompositeFactory_OptionsParseError=Failed to parse the input as valid JSON. Setting is not saved.
-JSLintValidatorPreferenceCompositeFactory_OptionsTitle=JSON Options
 JSParserValidatorPreferenceCompositeFactory_MissingSemicolons=Missing semicolons: 
 JSParserValidatorPreferenceCompositeFactory_OptionsGroup=Options
