+++ bundles/com.aptana.js.core/OSGI-INF/l10n/bundle.properties
@@ -7,8 +7,6 @@ content-type.sdocml.name = SDocML
 jsca.content-type.name = JSCA
 
 validator.language.js.name = JavaScript
-validator.jslint.name = JSLint Validator
-validator.mozilla.name = Mozilla JavaScript Validator
 validator.parser.name = JavaScript Syntax Validator
 validator.jsstyle.name = Javascript Style
 
@@ -20,8 +18,6 @@ sdocml.problem.name = SDocML Problem
 js.task.detector.name = JS Task Detector
 
 marker.js.name = JavaScript Syntax Problem
-marker.jslint.name= JSLint Problem
 marker.jsstyle.name= JavaScript Style Problem
-marker.mozilla.name= Mozilla JavaScript Problem
 extension-point.requireResolvers.name = Require Resolvers
 contributor.nodejs.name = NodeJS Source Contributor
