+++ bundles/com.aptana.js.core/build.properties
@@ -3,7 +3,6 @@ output.. = target/classes/
 bin.includes = META-INF/,\
                plugin.xml,\
                metadata/,\
-               fulljslint.js,\
                .,\
                OSGI-INF/l10n/bundle.properties,\
                .options
