+++ bundles/com.aptana.scripting/build.properties
@@ -23,3 +23,5 @@ src.includes = src/,\
                .project,\
                .settings/,\
                .classpath
+qualifier = 1556809251
+qualifier = 1556809251
