+++ features/com.aptana.feature/feature.xml
@@ -464,13 +464,6 @@
          unpack="false"/>
 
    <plugin
-         id="org.mozilla.javascript"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
-   <plugin
          id="org.w3c.css"
          download-size="0"
          install-size="0"
